Very basic map shifting and generation
Shifts map each turn and generates a new row.

Some stubbed out variables to enable row generation to be parameterized
by the "level" of the map.

Still a lot of code from robots; I'll need to cut it out as I go, but
some of the code can be repurposed for collision detection / sensing,
etc. so I'll leave it in for now.