Previous fix (for #18) did not handle teleport.
Teleporting when on an obstacle now destroys the obstacle.

From the code:

"if the player previously teleported when on an obstacle, just destroy the obstacle. We can't put it back where it was because we don't know (x, y) for the player due to map shifting, and we can't draw it under us or we will collide with it twice!"