A broken version on the way to an upgraded version.
Trying to get the code updated so that the point sensors work in the code with the new framwork.

Rather than subclassing the DefaultGridPlayer, I set player.update_state to the pre-existing read_bot_state function, which seemed to do the same thing when I compared it to Al's code in TRON.

When I do that, however, and press "play", the game runs and completes without error, and I can see that the sensors are being updated, but the code does not respond to the values of the sensors and nothing appears in the game window. If I press "frame forward" a playfield will appear and it will manually advance through the frames, but play does not work. Show / hide debug table acts like it works, but nothing appears, so I can't see what's going on inside the program state.

Again, moves are not being set or are not being lost somehow; I can change the entire program to be 'move = west' and the bot just drives north until it crashes enough to die.

On the other hand, if I comment out 'self.player.update_state = self.read_bot_state' on line 135 of game.py, the game runs, and the display works, but the sensors are not updated. When commented out, the program works, so a program of 'move = west' does exactly what you'd expect.

What on earth is going on?

I've set DEBUG to true and added a pretty-printing function for bot_vars.